requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
pdfplumber==0.11.4
python-dotenv==1.0.1
